[LETS-285] passive tran server recovery analysis (#3251)
http://jira.cubrid.org/browse/LETS-285

Run recovery analysis to rebuild the starting transaction table snapshot upon start of passive transaction server.

Implementation
- create an initial transaction table image based on log record containing transaction table snapshot
- run analysis and update the transaction table image until the end of the log already available on the passive transaction server (aka: catch-up)
- at the end of the analysis, log recovery analysis does not reset the LSA on the passive transaction server because the PTS already has consistent such LSA copied from the page server
- at the end of the analysis, on PTS, the system transactions created during recovery analysis are cleared because they are not relevant to PTS - PTS does not execute undo and postpones like the regular recovery

Other:
- fixed some compiler minor warnings

+++ src/executables/master.c
@@ -604,7 +604,7 @@ css_register_new_server2 (CSS_CONN_ENTRY * conn, unsigned short rid)
 		{
 		  entry->port_id = ntohl (buffer);
 		  entry->server_type = type;
-		  int length = (int) strlen (server_name.c_str ()) + 1;
+		  unsigned length = strlen (server_name.c_str ()) + 1;
 		  /* read server version_string, env_var, pid */
 		  if (length < server_name.length ())
 		    {
